# Improve performance of safe_load function