Enhance the performance of the `safe_load` function by optimizing data handling and device management, ensuring efficient loading of tensor data from various sources.
This improves the state dict loading too.

benchmark:
```bash
07:34:29 jlucaso@jlucaso jsgrad ±|main ✗|→ /usr/bin/time deno run -A cli/whisper.ts --input ./output.wav
Loading model safe_load: 1426ms
Loading model Whisper.init: 1947ms
Loading state dict: 100.0% |████████████████| 168/168 | 4.83>0.00s 34.78/s
Loading model load_state_dict: 4833ms
15.76user 1.17system 0:08.65elapsed 195%CPU (0avgtext+0avgdata 2223320maxresident)k
0inputs+216outputs (0major+493899minor)pagefaults 0swaps
```

```bash
07:34:50 jlucaso@jlucaso jsgrad ±|perf-safe-load ✗|→ /usr/bin/time deno run -A cli/whisper.ts --input ./output.wav
Loading model safe_load: 188ms
Loading model Whisper.init: 1905ms
Loading state dict: 100.0% |████████████████| 168/168 | 0.35>0.00s 481.38/s
Loading model load_state_dict: 350ms
5.02user 0.65system 0:02.88elapsed 196%CPU (0avgtext+0avgdata 1421552maxresident)k
0inputs+216outputs (0major+278720minor)pagefaults 0swaps
```