@karelnagel 2025-04-05
Tested this on my stable-diffusion branch and it is indeed much much faster. I guess tinygrad made it like it was so that they can use disk device mmap for larger files, but since jsgrad doesn't support mmap yet I think we can use something like this. 

plus stable-diffusion with WEBGPU threw `dtypes.long not supported` and this solved that as well. 

`source_device` doesn't seem to be used for anything, I will remove it in SD branch and merge this now to get that one working. 

Thanks!